+++ simaple/data/baseline/resources/item_name_alias.json
