+++ simaple/data/baseline/spec/Legendary.yaml
