+++ simaple/data/baseline/spec/LegendaryHalf.yaml
